Add GPMDP Compatibility
This checks to see if the Now Playing filename is playback.json and if it is reads the JSON file according to the Google Play Music Desktop Player JSON structure.  Kinda hacky and would be better done with a checkbox asking if it's GPMDP but I was mainly doing this for my own purposes, feel free to add it in though! [GPMDP](https://github.com/MarshallOfSound/Google-Play-Music-Desktop-Player-UNOFFICIAL-)