removing TZ adjustment for gallery files
MakeGallery.php was adjusting the filetime based on the timezone offset from GMT.  This caused the check on line 171 that verified if the screenshot took place after the system was enterred to fail.

I removed the references to $filetime_o, and set $filetime to the modification time of the image file.